python/weather/pyproject.toml
[project]
name = "weather-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "optimizely-opal-opal-tools-sdk>=0.1.1.dev0",
]
